Fixed assumption the abbreviations are always 3 characters
Removed the division by 4 (3 chars and an and of string), and replaced
it with count the  number of 0's up to the  abbrind. This corrects the
index for  Europe/Amsterdam. In that  case the offset would  become 7,
one  more than  was  defined  in the  array.  This  because there  are
multiple non 3 character zones for that timezone.